% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmix_loglik.R
\name{mbmix_loglik}
\alias{mbmix_loglik}
\title{log likelihood function for microbiome mixture}
\usage{
mbmix_loglik(tab, p, e = 0)
}
\arguments{
\item{tab}{Dataset of read counts as 3d array of size 3x3x2, genotype in first sample x genotype in second sample x allele in read.}

\item{p}{Contaminant probability (proportion of mixture coming from the second sample).}

\item{e}{Sequencing error rate.}
}
\value{
The log likelihood evaluated at \code{p} and \code{e}.
}
\description{
Calculate log likelihood function for microbiome sample mixture model at particular values of \code{p} and \code{e}.
}
\examples{
data(mbmixdata)
mbmix_loglik(mbmixdata, p=0.74, e=0.002)

}
