% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmixdata.R
\docType{data}
\name{mbmixdata}
\alias{mbmixdata}
\title{Example dataset for mbmixture package}
\format{
Dataset of read counts as 3d array of size 3x3x2, genotype in first sample x genotype in second sample x allele in read.
}
\usage{
data(mbmixdata)
}
\description{
Example dataset for mbmixture package.
}
\examples{
data(mbmixdata)
mle_pe(mbmixdata)
}
\keyword{datasets}
