% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmixture-package.R
\docType{package}
\name{mbmixture-package}
\alias{mbmixture}
\alias{mbmixture-package}
\title{mbmixture: Microbiome Mixture Analysis}
\description{
Evaluate whether a microbiome sample is a mixture of two samples, by fitting a model for the number of read counts as a function of single nucleotide polymorphism (SNP) allele and the genotypes of two potential source samples. Lobo et al. (2021) \doi{10.1093/g3journal/jkab308}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/mbmixture}
  \item Report bugs at \url{https://github.com/kbroman/mbmixture/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
